rag_pipeline/rag_pipeline/config.py
import os
from dotenv import load_dotenv

load_dotenv()

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
PINECONE_API_KEY = os.getenv("PINECONE_API_KEY")
INDEX_NAME = os.getenv("INDEX_NAME", "rag-chunking-eval")
INDEX_NAME_PIPELINE = os.getenv("INDEX_NAME_PIPELINE", "semantic-rag")

if not OPENAI_API_KEY or not PINECONE_API_KEY:
    raise ValueError("Missing API keys. Please check your .env file.")